show data in charts

+++ app/src/main/java/com/example/ems/repository/ActivePowerDTO.kt
@@ -4,9 +4,9 @@ import com.google.gson.annotations.SerializedName
 import java.util.*
 
 data class ActivePowerDTO(
-  @SerializedName("building_active_power") val building: Float,
-  @SerializedName("grid_active_power") val grid: Float,
-  @SerializedName("pv_active_power") val pv: Float,
-  @SerializedName("quasars_active_power") val quasars: Float,
+  @SerializedName("building_active_power") val building: Double,
+  @SerializedName("grid_active_power") val grid: Double,
+  @SerializedName("pv_active_power") val pv: Double,
+  @SerializedName("quasars_active_power") val quasars: Double,
   @SerializedName("timestamp") val date: Date
 )
